move glyph to text component
