+++ sugarloaf/src/components/text/glyph/cache/geometry.rs
