+++ sugarloaf/src/components/text/glyph/cache/mod.rs
@@ -1025,12 +1025,12 @@ fn draw_glyph(
 
 #[cfg(test)]
 mod test {
-    use super::*;
-    use crate::glyph::layout::*;
+    use crate::components::text::glyph::cache::*;
+    use crate::components::text::glyph::layout::*;
     use approx::*;
 
     const FONT: &[u8] =
-        include_bytes!("../../../resources/test-fonts/WenQuanYiMicroHei.ttf");
+        include_bytes!("../../../../../resources/test-fonts/WenQuanYiMicroHei.ttf");
 
     #[test]
     fn cache_test() {
@@ -1053,7 +1053,7 @@ mod test {
         for &(text, scale) in &strings {
             println!("Caching {:?}", (text, scale));
 
-            let glyphs = crate::glyph::layout::Layout::default_single_line()
+            let glyphs = crate::components::text::glyph::Layout::default_single_line()
                 .calculate_glyphs(
                     &[&font],
                     &SectionGeometry::default(),
@@ -1205,7 +1205,7 @@ mod test {
             .position_tolerance(0.1)
             .build();
 
-        let glyphs = crate::glyph::layout::Layout::default_single_line()
+        let glyphs = crate::components::text::glyph::Layout::default_single_line()
             .calculate_glyphs(
                 fontmap,
                 &SectionGeometry::default(),
@@ -1220,7 +1220,7 @@ mod test {
         }
         assert_eq!(cache.cache_queued(fontmap, |_, _| {}), Ok(CachedBy::Adding));
 
-        let glyphs = crate::glyph::layout::Layout::default_single_line()
+        let glyphs = crate::components::text::glyph::Layout::default_single_line()
             .calculate_glyphs(
                 fontmap,
                 &SectionGeometry::default(),
