+++ sugarloaf/src/lib.rs
@@ -2,10 +2,8 @@ pub mod components;
 pub mod content;
 pub mod context;
 pub mod font;
-pub mod glyph;
 pub mod layout;
 mod sugarloaf;
-pub mod tools;
 
 pub use crate::content::{Content, ContentBuilder};
 pub use crate::sugarloaf::{
